requirements.txt
groq
Pillow
google-generativeai
opencv-python
pyperclip
pyaudio
faster-whisper
speechrecognition
pyttsx3
requests
python-dotenv